+++ core/src/test/java/com/graphhopper/routing/util/parsers/OSMHazmatTunnelParserTest.java
@@ -30,31 +30,31 @@ public class OSMHazmatTunnelParserTest {
         IntsRef intsRef = em.createEdgeFlags();
         ReaderWay readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:adr_tunnel_cat", "A");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.A, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:adr_tunnel_cat", "B");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.B, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:adr_tunnel_cat", "C");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.C, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:adr_tunnel_cat", "D");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.D, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:adr_tunnel_cat", "E");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.E, hazTunnelEnc.getEnum(false, intsRef));
     }
 
@@ -63,31 +63,31 @@ public class OSMHazmatTunnelParserTest {
         IntsRef intsRef = em.createEdgeFlags();
         ReaderWay readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:tunnel_cat", "A");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.A, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:tunnel_cat", "B");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.B, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:tunnel_cat", "C");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.C, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:tunnel_cat", "D");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.D, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:tunnel_cat", "E");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.E, hazTunnelEnc.getEnum(false, intsRef));
     }
 
@@ -97,35 +97,35 @@ public class OSMHazmatTunnelParserTest {
         ReaderWay readerWay = new ReaderWay(1);
         readerWay.setTag("tunnel", "yes");
         readerWay.setTag("hazmat:A", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.A, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("tunnel", "yes");
         readerWay.setTag("hazmat:B", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.B, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("tunnel", "yes");
         readerWay.setTag("hazmat:C", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.C, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("tunnel", "yes");
         readerWay.setTag("hazmat:D", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.D, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
         readerWay = new ReaderWay(1);
         readerWay.setTag("tunnel", "yes");
         readerWay.setTag("hazmat:E", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.E, hazTunnelEnc.getEnum(false, intsRef));
     }
 
@@ -139,7 +139,7 @@ public class OSMHazmatTunnelParserTest {
         readerWay.setTag("hazmat:C", "no");
         readerWay.setTag("hazmat:D", "no");
         readerWay.setTag("hazmat:E", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.E, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
@@ -148,7 +148,7 @@ public class OSMHazmatTunnelParserTest {
         readerWay.setTag("hazmat:A", "no");
         readerWay.setTag("hazmat:B", "no");
         readerWay.setTag("hazmat:C", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.C, hazTunnelEnc.getEnum(false, intsRef));
 
         intsRef = em.createEdgeFlags();
@@ -156,7 +156,7 @@ public class OSMHazmatTunnelParserTest {
         readerWay.setTag("tunnel", "yes");
         readerWay.setTag("hazmat:B", "no");
         readerWay.setTag("hazmat:E", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.E, hazTunnelEnc.getEnum(false, intsRef));
     }
 
@@ -165,7 +165,7 @@ public class OSMHazmatTunnelParserTest {
         IntsRef intsRef = em.createEdgeFlags();
         ReaderWay readerWay = new ReaderWay(1);
         readerWay.setTag("hazmat:B", "no");
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.A, hazTunnelEnc.getEnum(false, intsRef));
     }
 
@@ -173,7 +173,7 @@ public class OSMHazmatTunnelParserTest {
     public void testNoNPE() {
         ReaderWay readerWay = new ReaderWay(1);
         IntsRef intsRef = em.createEdgeFlags();
-        parser.handleWayTags(intsRef, readerWay, false, relFlags);
+        parser.handleWayTags(intsRef, readerWay, relFlags);
         assertEquals(HazmatTunnel.A, hazTunnelEnc.getEnum(false, intsRef));
     }
 }