+++ core/src/test/java/com/graphhopper/routing/util/parsers/OSMMaxWeightParserTest.java
@@ -29,13 +29,13 @@ public class OSMMaxWeightParserTest {
         IntsRef intsRef = em.createEdgeFlags();
         readerWay.setTag("highway", "primary");
         readerWay.setTag("maxweight", "5");
-        parser.handleWayTags(intsRef, readerWay, false, em.createRelationFlags());
+        parser.handleWayTags(intsRef, readerWay, em.createRelationFlags());
         assertEquals(5.0, mwEnc.getDecimal(false, intsRef), .01);
 
         // if value is beyond the maximum then do not use infinity instead fallback to more restrictive maximum
         intsRef = em.createEdgeFlags();
         readerWay.setTag("maxweight", "50");
-        parser.handleWayTags(intsRef, readerWay, false, em.createRelationFlags());
+        parser.handleWayTags(intsRef, readerWay, em.createRelationFlags());
         assertEquals(mwEnc.getMaxDecimal(), mwEnc.getDecimal(false, intsRef), .01);
     }
 }